refactor(core): promote ModelRef to core/ with a re-export shim
ModelRef is a weightless *contract*, not a matcher: a stdlib-only frozen
dataclass (base, adapter) that round-trips through Resolver.save/load as
plain JSON. Move it to core/ beside the other contracts; leave a pure
re-export shim at the old matchers/ path.

Sequencing: a later wave moves finetune.py into training/ while a different
wave rewrites finetune.py's model_ref import -- same line, two branches.
Landing the promotion + shim now decouples them.

Zero behaviour change: both import paths resolve to the same objects.
